botched commit :)

+++ delete.go
@@ -1,5 +1,6 @@
 package main
 
+import "fmt"
 import "os"
 
 // syncdel deletes a list of files synchronously.
@@ -10,7 +11,7 @@ func syncdel(files *[]string) {
 		currentTaskType = 4
 		currentFile = path
 
-		if !dryrun {
+		if !isnodelete(path) && !dryrun {
 			err = os.RemoveAll(path)
 
 			if err != nil {
@@ -18,4 +19,15 @@ func syncdel(files *[]string) {
 			}
 		}
 	}
+}
+
+func isnodelete(path string) bool {
+	fmt.Println("\n\n\n\ninod:" + path + "\n\n\n\n\n")
+	for _, p := range nodelete {
+		if p == path {
+			return true
+		}
+	}
+
+	return false
 }
